UPDATE 1.2
se agregan funciones para
nodo_interno(nodo_inter_liberar,nodo_inter__cargar) y pequeños arreglos
en condición de identificar cabezal del archivo

+++ libs/contacto.h
@@ -5,7 +5,7 @@ typedef struct contacto
 	int clave;
 	char nombre[LIMIT_NAMES];
 	char apellido[LIMIT_NAMES];
-	unsigned int telefono;
+	unsigned long telefono;
 } Contacto;
 
 Contacto* contacto_Crear();
